+++ src/anomalib/data/base/depth.py
@@ -9,8 +9,8 @@ import numpy as np
 import torch
 
 from anomalib.data.base.dataset import AnomalibDataset
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import masks_to_boxes, read_depth_image, read_image
+from anomalib.utils.types import TaskType
 
 
 class AnomalibDepthDataset(AnomalibDataset, ABC):
