+++ src/anomalib/data/image/btech.py
@@ -21,7 +21,6 @@ from pandas.core.frame import DataFrame
 from tqdm import tqdm
 
 from anomalib.data.base import AnomalibDataModule, AnomalibDataset
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
@@ -32,6 +31,7 @@ from anomalib.data.utils import (
     download_and_extract,
     get_transforms,
 )
+from anomalib.utils.types import TaskType
 
 logger = logging.getLogger(__name__)
