+++ src/anomalib/data/image/folder.py
@@ -14,7 +14,6 @@ import albumentations as A  # noqa: N812
 from pandas import DataFrame
 
 from anomalib.data.base import AnomalibDataModule, AnomalibDataset
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DirType,
     InputNormalizationMethod,
@@ -25,6 +24,7 @@ from anomalib.data.utils import (
     get_transforms,
 )
 from anomalib.data.utils.path import _prepare_files_labels, _resolve_path
+from anomalib.utils.types import TaskType
 
 
 def make_folder_dataset(
