+++ src/anomalib/data/image/visa.py
@@ -31,7 +31,6 @@ import albumentations as A  # noqa: N812
 import cv2
 
 from anomalib.data.base import AnomalibDataModule, AnomalibDataset
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
@@ -41,6 +40,7 @@ from anomalib.data.utils import (
     download_and_extract,
     get_transforms,
 )
+from anomalib.utils.types import TaskType
 
 from .mvtec import make_mvtec_dataset
