+++ src/anomalib/data/utils/synthetic.py
@@ -21,8 +21,8 @@ from albumentations.pytorch import ToTensorV2
 from pandas import DataFrame, Series
 
 from anomalib.data.base.dataset import AnomalibDataset
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import Augmenter, Split, read_image
+from anomalib.utils.types import TaskType
 
 logger = logging.getLogger(__name__)
