+++ src/anomalib/data/video/avenue.py
@@ -28,7 +28,6 @@ from pandas import DataFrame
 
 from anomalib.data.base import AnomalibVideoDataModule, AnomalibVideoDataset
 from anomalib.data.base.video import VideoTargetFrame
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
@@ -38,6 +37,7 @@ from anomalib.data.utils import (
     get_transforms,
 )
 from anomalib.data.utils.video import ClipsIndexer
+from anomalib.utils.types import TaskType
 
 if TYPE_CHECKING:
     from collections.abc import Callable
