+++ src/anomalib/data/video/shanghaitech.py
@@ -27,7 +27,6 @@ from pandas import DataFrame
 
 from anomalib.data.base import AnomalibVideoDataModule, AnomalibVideoDataset
 from anomalib.data.base.video import VideoTargetFrame
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
@@ -38,6 +37,7 @@ from anomalib.data.utils import (
     read_image,
 )
 from anomalib.data.utils.video import ClipsIndexer, convert_video
+from anomalib.utils.types import TaskType
 
 logger = logging.getLogger(__name__)
