+++ src/anomalib/data/video/ucsd_ped.py
@@ -17,7 +17,6 @@ from pandas import DataFrame
 
 from anomalib.data.base import AnomalibVideoDataModule, AnomalibVideoDataset
 from anomalib.data.base.video import VideoTargetFrame
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
@@ -28,6 +27,7 @@ from anomalib.data.utils import (
     read_image,
 )
 from anomalib.data.utils.video import ClipsIndexer
+from anomalib.utils.types import TaskType
 
 if TYPE_CHECKING:
     from collections.abc import Callable
