+++ src/anomalib/deploy/export.py
@@ -14,9 +14,10 @@ import albumentations as A  # noqa: N812
 import numpy as np
 import torch
 
-from anomalib.data import AnomalibDataModule, AnomalibDataset, TaskType
+from anomalib.data import AnomalibDataModule, AnomalibDataset
 from anomalib.models.components import AnomalyModule
 from anomalib.utils.exceptions import try_import
+from anomalib.utils.types import TaskType
 
 if TYPE_CHECKING:
     from torch.types import Number
