+++ src/anomalib/deploy/inferencers/openvino_inferencer.py
@@ -14,7 +14,7 @@ import cv2
 import numpy as np
 from omegaconf import DictConfig
 
-from anomalib.data import TaskType
+from anomalib.utils.types import TaskType
 
 from .base_inferencer import Inferencer
