+++ src/anomalib/engine/engine.py
@@ -15,7 +15,7 @@ from lightning.pytorch.trainer.connectors.callback_connector import _CallbackCon
 from lightning.pytorch.utilities.types import _EVALUATE_OUTPUT, _PREDICT_OUTPUT, EVAL_DATALOADERS, TRAIN_DATALOADERS
 from omegaconf import DictConfig, ListConfig
 
-from anomalib.data import AnomalibDataModule, AnomalibDataset, TaskType
+from anomalib.data import AnomalibDataModule, AnomalibDataset
 from anomalib.deploy.export import ExportMode, export_to_onnx, export_to_openvino, export_to_torch
 from anomalib.models import AnomalyModule
 from anomalib.post_processing import NormalizationMethod
@@ -25,6 +25,7 @@ from anomalib.utils.callbacks.normalization import get_normalization_callback
 from anomalib.utils.callbacks.post_processor import _PostProcessorCallback
 from anomalib.utils.callbacks.thresholding import _ThresholdCallback
 from anomalib.utils.metrics.threshold import BaseThreshold
+from anomalib.utils.types import TaskType
 
 logger = logging.getLogger(__name__)
