+++ src/anomalib/post_processing/visualizer.py
@@ -15,7 +15,6 @@ import matplotlib.pyplot as plt
 import numpy as np
 from skimage.segmentation import mark_boundaries
 
-from anomalib.data import TaskType
 from anomalib.data.utils import read_image
 from anomalib.post_processing.post_process import (
     add_anomalous_label,
@@ -23,6 +22,7 @@ from anomalib.post_processing.post_process import (
     draw_boxes,
     superimpose_anomaly_map,
 )
+from anomalib.utils.types import TaskType
 
 
 @dataclass
