+++ src/anomalib/utils/callbacks/metrics.py
@@ -12,9 +12,9 @@ import torch
 from lightning.pytorch import Callback, Trainer
 from lightning.pytorch.utilities.types import STEP_OUTPUT
 
-from anomalib.data import TaskType
 from anomalib.models import AnomalyModule
 from anomalib.utils.metrics import AnomalibMetricCollection, create_metric_collection
+from anomalib.utils.types import TaskType
 
 logger = logging.getLogger(__name__)
