+++ tests/integration/cli/test_cli.py
@@ -13,9 +13,9 @@ from pathlib import Path
 import pytest
 import torch
 
-from anomalib.data import TaskType
 from anomalib.models import AnomalyModule
 from anomalib.utils.cli import AnomalibCLI
+from anomalib.utils.types import TaskType
 
 
 @pytest.fixture(scope="module")
