+++ tests/legacy/helpers/model.py
@@ -12,12 +12,13 @@ from lightning.pytorch.callbacks import ModelCheckpoint
 from omegaconf import DictConfig, ListConfig
 
 from anomalib.config import _update_nncf_config
-from anomalib.data import TaskType, get_datamodule
+from anomalib.data import get_datamodule
 from anomalib.engine import Engine
 from anomalib.models import get_model
 from anomalib.models.components import AnomalyModule
 from anomalib.utils.callbacks import get_callbacks
 from anomalib.utils.callbacks.visualizer import BaseVisualizerCallback
+from anomalib.utils.types import TaskType
 
 
 def model_load_test(config: Union[DictConfig, ListConfig], datamodule: LightningDataModule, results: Dict):
