+++ tests/legacy/pre_merge/tools/test_openvino_entrypoint.py
@@ -11,10 +11,10 @@ from pathlib import Path
 
 import pytest
 
-from anomalib.data import TaskType
 from anomalib.deploy import export_to_openvino
 from anomalib.models import Padim
 from anomalib.utils.metrics.threshold import F1AdaptiveThreshold
+from anomalib.utils.types import TaskType
 
 sys.path.append("tools/inference")
