+++ tests/unit/data/image/test_mvtec.py
@@ -7,7 +7,8 @@ from pathlib import Path
 
 import pytest
 
-from anomalib.data import MVTec, TaskType
+from anomalib.data import MVTec
+from anomalib.utils.types import TaskType
 from tests.unit.data.base.image import _TestAnomalibImageDatamodule
