+++ tests/unit/data/image/test_visa.py
@@ -7,7 +7,8 @@ from pathlib import Path
 
 import pytest
 
-from anomalib.data import TaskType, Visa
+from anomalib.data import Visa
+from anomalib.utils.types import TaskType
 from tests.unit.data.base.image import _TestAnomalibImageDatamodule
