+++ tests/unit/data/test_synthetic.py
@@ -8,10 +8,10 @@ from pathlib import Path
 
 import pytest
 
-from anomalib.data import TaskType
-from anomalib.data.folder import FolderDataset
-from anomalib.data.synthetic import SyntheticAnomalyDataset
+from anomalib.data.image.folder import FolderDataset
 from anomalib.data.utils import get_transforms
+from anomalib.data.utils.synthetic import SyntheticAnomalyDataset
+from anomalib.utils.types import TaskType
 
 
 @pytest.fixture(scope="module")
