+++ tests/unit/data/video/test_shanghaitech.py
@@ -8,7 +8,8 @@ from pathlib import Path
 
 import pytest
 
-from anomalib.data import ShanghaiTech, TaskType
+from anomalib.data import ShanghaiTech
+from anomalib.utils.types import TaskType
 from tests.unit.data.base.video import _TestAnomalibVideoDatamodule
