+++ tests/unit/post_processing/test_visualizer.py
@@ -11,11 +11,12 @@ import pytest
 from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
 from torch.utils.data import DataLoader
 
-from anomalib.data import MVTec, TaskType
+from anomalib.data import MVTec
 from anomalib.data.inference import InferenceDataset
 from anomalib.engine import Engine
 from anomalib.models import get_model
 from anomalib.post_processing.visualizer import ImageGrid
+from anomalib.utils.types import TaskType
 
 
 def test_visualize_fully_defected_masks() -> None:
